+++ swig/RecurrenceState.swig.h
@@ -1,18 +1,6 @@
 /*
+ * SPDX-License-Identifier: AGPL-3.0-only
  * Copyright 2005 - 2016 Zarafa and its licensors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
 /* Copy of the full common/RecurrenceState.h file, but without the
@@ -111,8 +99,8 @@ public:
 
 class RecurrenceState {
 public:
-	HRESULT ParseBlob(const char *lpData, unsigned int ulLen, ULONG ulFlags);
-    HRESULT GetBlob(char **lpData, unsigned int *lpulLen, void *base = NULL);
+	HRESULT ParseBlob(const char *lpData, size_t ulLen, ULONG ulFlags);
+	HRESULT GetBlob(char **lpData, size_t *lpulLen, void *base = NULL);
 
 	/* not inlined for swig */
